changing a bit how seeds work to resolve potential conflicts when setting up a new DB:
