+++ app/database/seed.server.ts
@@ -1,14 +1,81 @@
 /* eslint-disable no-console */
-import { PrismaClient } from "@prisma/client";
+import type { Role } from "@prisma/client";
+import { PrismaClient, Roles } from "@prisma/client";
 import { createClient } from "@supabase/supabase-js";
 
 import { createUser } from "~/modules/user";
 import { ShelfStackError } from "~/utils/error";
-import {
-  createAdminRole,
-  createUserRole,
-} from "./manual-migrations/master-data/seed-role.server";
+
 import { SUPABASE_SERVICE_ROLE, SUPABASE_URL } from "../utils/env";
+
+export const createUserRole = async () => {
+  const existingRole = await prisma.role.findFirst({
+    where: {
+      name: Roles["USER"],
+    },
+  });
+
+  if (existingRole) return null;
+
+  return await prisma.role.create({
+    data: {
+      name: Roles["USER"],
+    },
+  });
+};
+
+export const createAdminRole = async () => {
+  const existingRole = await prisma.role.findFirst({
+    where: {
+      name: Roles["ADMIN"],
+    },
+  });
+
+  if (existingRole) return null;
+
+  return await prisma.role.create({
+    data: {
+      name: Roles["ADMIN"],
+    },
+  });
+};
+
+export const addUserRoleToAllExistingUsers = async () => {
+  const allUsers = await prisma.user.findMany({
+    include: {
+      roles: true,
+    },
+  });
+
+  const userRole = (await prisma.role.findFirst({
+    where: {
+      name: Roles["USER"],
+    },
+  })) as Role;
+
+  allUsers.map(async (user) => {
+    if (
+      user.roles?.some(
+        (role) => role.name === Roles["USER"] || role.name === Roles["ADMIN"]
+      )
+    )
+      return;
+    return await prisma.user.update({
+      where: {
+        id: user.id,
+      },
+      data: {
+        roles: {
+          connect: {
+            id: userRole.id,
+          },
+        },
+      },
+    });
+  });
+
+  return allUsers;
+};
 
 const supabaseAdmin = createClient(SUPABASE_URL, SUPABASE_SERVICE_ROLE, {
   auth: {
@@ -81,6 +148,8 @@ async function seed() {
     if (!user) {
       throw new ShelfStackError({ message: "Unable to create user" });
     }
+
+    await addUserRoleToAllExistingUsers();
 
     console.log(`Database has been seeded. 🌱\n`);
     console.log(
